API for learning_pi.jl

Index

LearningPi.Corpus

LearningPi.Graphormer

LearningPi.LayerNorm

LearningPi.LayerNorm

LearningPi.RMSNorm

LearningPi.abstract_dataset

LearningPi.abstract_deviation

LearningPi.abstract_example

LearningPi.abstract_features_matrix

LearningPi.cr_deviation

LearningPi.cr_features_matrix

LearningPi.featuresCWL

LearningPi.featuresGA

LearningPi.featuresMCND

LearningPi.labels

LearningPi.labelsCWL

LearningPi.labelsGA

LearningPi.learningType

LearningPi.loss_hinge

LearningPi.loss_hinge

LearningPi.loss_mse

LearningPi.lr_features_matrix

LearningPi.without_cr_features_matrix

LearningPi.zero_deviation

ChainRulesCore.rrule

ChainRulesCore.rrule

ChainRulesCore.rrule

LearningPi.compareWithBests

LearningPi.createCorpus

LearningPi.createDataSet

LearningPi.createDataSet

LearningPi.createEmptyDataset

LearningPi.createEmptyDataset

LearningPi.createKfold

LearningPi.createLabels

LearningPi.createLabels

LearningPi.createLabels

LearningPi.create_example

LearningPi.create_example

LearningPi.create_features

LearningPi.create_features

LearningPi.create_features

LearningPi.create_loss

LearningPi.create_model

LearningPi.create_rms

LearningPi.dataLoader

LearningPi.dataLoader

LearningPi.dataLoader

LearningPi.deviationFrom

LearningPi.deviationFrom

LearningPi.featuresExtraction

LearningPi.featuresExtraction

LearningPi.featuresExtraction

LearningPi.featuresExtraction

LearningPi.features_matrix

LearningPi.features_matrix

LearningPi.features_matrix

LearningPi.gap

LearningPi.gap_closure

LearningPi.get_cr_features

LearningPi.get_cr_features

LearningPi.get_cr_features

LearningPi.get_device

LearningPi.get_model

LearningPi.get_parameters

LearningPi.get_λ

LearningPi.gradient_lsp

LearningPi.gradient_lsp

LearningPi.gradient_lsp

LearningPi.prediction

LearningPi.printBests

LearningPi.printMetrics

LearningPi.print_best_models

LearningPi.print_json

LearningPi.print_json

LearningPi.print_json

LearningPi.read_labels

LearningPi.read_labels

LearningPi.read_labels

LearningPi.saveHP

LearningPi.saveHP

LearningPi.sizeFeatures

LearningPi.size_features_constraint

LearningPi.size_features_constraint

LearningPi.size_features_constraint

LearningPi.size_features_variable

LearningPi.size_features_variable

LearningPi.size_features_variable

LearningPi.sub_problem_value

LearningPi.sub_problem_value

LearningPi.testAndPrint

LearningPi.train

LearningPi.validation

Public

LearningPi.createCorpus — Function

createCorpus(featType::learningType, folder::String, maxInstance::Vector{Int64}=[-1, -1, -1],seed::Int64=0;factory::abstractInstanceFactory=cpuMCNDinstanceFactory(),pTrain::Float32=0.8,pVal::Float32=0.1)

Arguments:

featType : the type of the features instance.

folder : the path to the directory that contains the json files that defines the instances (and the associated features and labels).

maxInstance : a vector with three components. that say how many instances take for the training/validation/test set.

seed : a random seed used to select which instaces consider in the training/validation/test sets.

factory: type of instance.

pTrain : The percentage of training instances in the provided folder.

pVal : The percentage of validation instances in the provided folder.

Create a Corpus, that is a structure with three DataSet field for training,validation and test dataset. Note: The percentage for the test set will be 1-pTrain-pVal. It is important to select the percentage of training and validation in such a way that pTrain+pVal<1 and both will be non-negative.

LearningPi.createDataSet — Function

function createDataSet(lt::learningGNN, directory::String, maxInstance::Int64=-1, factory::abstractInstanceFactory)

#Arguments:
	- `lt`:learning type, it should be a sub-type learningGNN 
	- `directory`: the path to the directory containing instances
	- `maxInstance`: maximum instance number 
	- `factory`: instance factory, generic sub-type of abstractInstanceFactory

Create and returns a dataset for the provided learning type `lt`, considering `maxInst` instances of the factory `factory`, contained in `directory`

LearningPi.createDataSet — Function

createDataSet(lt::learningType, directory, maxInstance::Int64=-1,factory::abstractInstanceFactory=cpuMCNDinstanceFactory())

Arguments:

lt: the learning type.

directory: a list of paths to the instances.

maxInstance: the maximum number of instances that we wat consider in the provided directory. By default is equal to -1, that means consider all the instances in the directory.

factory: type of instance, the possibilities in this moment are cpuMCNDinstanceFactory() (that is Multi-Commodity Network-Design instances) and cpuCWLinstanceFactory() (for the Bin Packing instances).

Create the dataset for the provided (general) learning type. return a dataSet structure of a proper type.

LearningPi.createKfold — Function

createKfold(featType::learningType, folder::String, maxInstance::Vector{Int64}=[-1, -1, -1],seed::Int64=0;factory::abstractInstanceFactory=cpuMCNDinstanceFactory(),k::Int64=1)

Arguments:

featType : the type of the features instance.

folder : the path to the directory that contains the json files that defines the instances (and the associated features and labels).

maxInstance : a vector with three components. that say how many instances take for the training/validation/test set.

seed : a random seed used to select which instaces consider in the training/validation/test sets.

factory: type of instance.

k: the fold that we want select as test set. Note: 1 <= k <= 10.

Create a Corpus, that is a struct with three DataSet field for training/validation/test set

LearningPi.createLabels — Method

createLabels(π, x, y, LRarcs, objLR, ins::abstractInstanceMCND)

Arguments:

π: a (optimal) Lagrangian multipliers vector

x: the flow variables in the Lagrangian sub-problem, obtained afer the resolution of the sub-problem with multipliers π

y: the design variables in the Lagrangian sub-problem, obtained afer the resolution of the sub-problem with multipliers π

LRarcs: a vector containign the bounds for each edge of the Lagrangian sub-problem considering π as Lagrangian multipleirs vector

objLR: the bound of the Lagrangian sub-problem considering π as Lagrangian multipleirs vector

ins: the instance structure (standard instance formulation, without regularization).

For instance (not normalized), nothing to do, just return a proper label structure.

LearningPi.createLabels — Method

function createLabels(π, x, y, objLR, ins::instanceCWL)

#Arguments:

-π: optimal lagrangian multipliers vector. -x: primal solution of the Knapsack Lagrangian Relaxation associated to the variables that associate one items to a pack (using the optimal Lagrangian multipliers). -y: primal solution of the Knapsack Lagrangian Relaxation associated to the variables say if we use or not a pack (using the optimal Lagrangian multipliers). -objLR: optimal value of the Lagrangian Dual

ins: instance object, it should be of type sub-type of instanceCWL

Given all the fields construct a label structure for the Bin Packing Problem.

LearningPi.createLabels — Method

function createLabels(π, x, objLR, ins::instanceGA)

#Arguments: -π: optimal lagrangian multipliers vector. -x: primal solution of the Knapsack Lagrangian Relaxation associated to the variables that associate one items to a pack (using the optimal Lagrangian multipliers). -objLR: optimal value of the Lagrangian Dual

Given all the fields construct a label structure for the Bin Packing Problem.

LearningPi.create_example — Method

create_example(lt::learningGNN, fileName::String, factory::abstractInstanceFactory)

#Arguments:

lt: learning Type, this function works for all the learning types that use a graph representation of the instance

fileName: path to the json that contains all the information to construct the learning sample starting from an instance, its features and the labels.

factory: instance factory, it works for all the factory

returns an gnnExample_instance with all the information useful for the training.

LearningPi.create_example — Method

create_example(lt::learningArc, fileName::String,factory::abstractInstanceFactory)

Arguments:

lt: learning type, it should be learningMLP.

fileName: the name of the file json that contains the informations about the instance, its features and its labels.

factory: type of instance (it handle both with the normalized and un-normalized instances).

Create a structure containing the instance, the extracted features and the associated labels.

LearningPi.create_loss — Function

createloss(::HingeLoss)

Arguments:

-:: loss parameters, it should be a structure of type HingeLoss.

return the loss correspondent to loss paramameters of type HingeLoss

LearningPi.create_model — Function

create_model(lType::learningMLP, in, h, out=1, a=relu)

Arguments:

lType: general learningType.

in: size of the input layer.

h: a vector with the same length as the desired number of hidden layers and each component say how many nodes we want in the correspondent hidden layer.

out: size of the output layer, by default is equal to one.

a: the activation function for the hidden layers, by default is relu.

This function creates a model for the provided learning type (in order to use this variant it should be learningArc). The model is a multi layer perceptron with in nodes in the first layer, length(h) hidden layers (the i-th layer has h[i] nodes) and out nodes in the output layer (by default 1). By default each hidden layer use a relu activation function (the input and output layers have no activation function).

LearningPi.dataLoader — Method

dataLoader(fileName::String, factory::cpuGAinstanceFactory)

Arguments

- `fileName` : a path to a json file that contains the data for the instance, features and labels
- `factory` : an instance factory for the Bin Packing problem

It reads the instance, the features and the labels from the json and returns three structures that contains all the informations.

LearningPi.dataLoader — Method

dataLoader(fileName::String, factory::MCNDnstanceFactory)

Arguments

- `fileName` : a path to a json file that contains the data for the instance, features and labels
- `factory` : an instance object (for the same instance as the features file)

It reads the instance, the features and the labels from the json and returns three structures that contains all the informations.

LearningPi.dataLoader — Method

dataLoader(fileName::String, factory::cpuCWLinstanceFactory)

Arguments

- `fileName` : a path to a json file that contains the data for the instance, features and labels
- `factory` : an instance factory for the Bin Packing problem

It reads the instance, the features and the labels from the json located in fileName and returns three structures that contains all the informations.

LearningPi.featuresExtraction — Method

featuresExtraction(featType::learningMLP, featObj, ins)

Vectorization function for the features when we consider a learningNodeDemand encoding.

Arguments:

featType features type

features a features matrix

nbFeatures the number of features

LearningPi.featuresExtraction — Method

function featuresExtraction(lt::learningType, featObj, ins::instanceGA, fmt::abstractfeaturesmatrix)

#Arguments:

lt: learnign type, it should be a sub-type of learningType

featObj: features object containing all the characteristics

ins: instance structure, it should instanceGA

fmt: features matrix type.

Returns the bipartite graph representation with the associated nodes-features matrix.

LearningPi.featuresExtraction — Method

function featuresExtraction(lt::learningGNN, featObj, ins::abstractInstanceMCND)

Arguments:

lt: learnign type, it should be a sub-type of learningGNN

featObj: features object containing all the characteristics

ins: instance structure, it should be a sub-type of abstractInstanceMCND

Returns the bipartite graph representation with the associated nodes-features matrix.

LearningPi.featuresExtraction — Method

function featuresExtraction(lt::learningType, featObj, ins::instanceCWL,fmt::abstractfeaturesmatrix)

#Arguments:

lt: learnign type, it should be a sub-type of learningType

featObj: features object containing all the characteristics

ins: instance structure, it should instanceCWL

fmt: features matrix type.

Returns the bipartite graph representation with the associated nodes-features matrix.

LearningPi.prediction — Method

prediction(nn, f,ins,lt::learningMLP)

Arguments:

-nn: neural network model. -f: features matrix. -ins: structure containing the instance informations. -lt: learning type (general).

LearningPi.train — Method

function train(maxEp::Int64, dS::Corpus, nn, opt::Optimiser, loss::abstractLoss; printEpoch::Int64=10, endString::String, lt::learningType, dt::abstract_deviation,seed::Int64,bs::Int64=1)

Arguments:

maxEp: the maximum number of epochs for the learning algorithm.

dS: the Corpus structure that contains the training, validation and test sets.

nn: the neural network model.

opt: the optimizer used for the training.

loss: a structure that contains the parameters α and β of the loss.

printEpoch: the number of epochs in which print the metrics of the training.

endString: the string used to memorize the output files as best models and tensorboard logs.

dt: deviation type, it could deviate from zero or the duals of the continuous relaxation.

lt: learning type

seed: random seed for the random generators

bs batch size

This function performs the learning with the provided inputs and save the best models in a bson file.

Private

LearningPi.Corpus — Type

Corpus structure contains the three datasets: training, validation and test set

LearningPi.Graphormer — Method

function (m::Graphormer)(x)

Arguments:

x: input of the NN model of type Graphormer

Forward computation of a Graphormer m, the output is the concatenation of all the multipliers predicted by the model

LearningPi.LayerNorm — Type

struct LayerNorm

#Fields:

eps: regularization parameter

d: size of the input of the normalization layer

Describe the Layer Normalization for the provided parameters

LearningPi.LayerNorm — Method

function (m::LayerNorm)(x)

Perform a Layer Normalization using x as input .

LearningPi.RMSNorm — Type

struct RMSNorm

#Fields:

eps: additive regularization parameter

sqrtd: multiplicative regularization parameter

Describe the RMS Normalization for the provided parameters

LearningPi.abstract_dataset — Type

Abstract type for the dataset

LearningPi.abstract_deviation — Type

abstract type abstract_deviation end

Abstract type to handle the deviation vector, i.e. the starting point from which our model produce an additive activation.

LearningPi.abstract_example — Type

Abstract type for an element of a dataset

LearningPi.abstract_features_matrix — Type

abstract type abstract_features_matrix

Abstract type for the construction of the nodes-features matrix associated to the bipartite graph representation of the instance.

LearningPi.cr_deviation — Type

struct cr_deviation end

Type to use as deviation vector (i.e. the starting point from which our model produce an additive activation) the dual variables associated to the relaxed constraints in the optimal solution of the Continuous Relaxation.

LearningPi.cr_features_matrix — Type

struct cr_features_matrix

with this choice the features matrix considers the informations related to the continuous relaxation of the instance.

LearningPi.featuresCWL — Type

Features structure for the Bin Packing instance. #Fields: -xCR: primal solution of the Linear Relaxation associated to the variables that associate one items to a pack. -yCR: primal solution of the Linear Relaxation associated to the variables say if we use or not a pack. -λ: dual solution of the Linear Relaxation associated to the packing constraints. -μ: dual solution of the Linear Relaxation associated to the packing constraints. -objCR: objective value of the Linear Relaxation. -xLR: primal solution of the Knapsack Lagrangian Relaxation associated to the variables that associate one items to a pack (using the dual variables λ of the linear relaxation). -yLR: primal solution of the Knapsack Lagrangian Relaxation associated to the variables say if we use or not a pack (using the dual variables λ of the linear relaxation). -objLR: objective value of the Knapsack Lagrangian Relaxation (using the dual variables λ of the linear relaxation).

LearningPi.featuresGA — Type

Features structure for the Bin Packing instance. #Fields: -xCR: primal solution of the Linear Relaxation associated to the variables that associate one items to a pack. -λ: dual solution of the Linear Relaxation associated to the packing constraints. -μ: dual solution of the Linear Relaxation associated to the packing constraints. -objCR: objective value of the Linear Relaxation. -xLR: primal solution of the Knapsack Lagrangian Relaxation associated to the variables that associate one items to a pack (using the dual variables λ of the linear relaxation). -objLR: objective value of the Knapsack Lagrangian Relaxation (using the dual variables λ of the linear relaxation).

LearningPi.featuresMCND — Type

Struct containing the information of the features for an instance.

Fields:

xCR: the value of the flow variables for the optimal solution of the linear relaxation.

yCR: the value of the decision variables for the optimal solution of the linear relaxation.

λ: the value of the dual variables associated to the flow constraints, for the optimal solution of the linear relaxation.

μ: the value of the dual variables associated to the capacity constraints for the optimal solution of the linear relaxation.

objCR: the objective value of the linear relaxation.

xLR: the value of the flow variables for the optimal solution of the sub-problem for the knapsack relaxation, considering as lagrangian multiers the vector λ.

yLR: the value of the design variables for the optimal solution of the sub-problem for the knapsack relaxation, considering as lagrangian multiers the vector λ.

LRarcs: the objective values, for each edge, of the optimal solution of the sub-problem for the knapsack relaxation, considering as lagrangian multiers the vector λ.

objLR: the objective value of the sub-problem for the knapsack relaxation, considering as lagrangian multiers the vector λ.

origins: a matrix of size K×V the cost of the shortest path from the origin to the current node with costs in an edge e: ins.r[k,e]+ins.f[e]/ins.c[e].

destinations: a matrix of size K×V the cost of the shortest path from the current node to the destination with costs in an edge e: ins.r[k,e]+ins.f[e]/ins.c[e].

distance: a matrix of size V×V with the distance in terms of number of edges for the shortest path from each two nodes.

LearningPi.labels — Type

Struct containing the information relative to the labels

Fields

π::Matrix{Float32}: matrix containing the gold Lagrangian multipliers. π[k, i] gives the values of the Lagrangian multiplier associated with demand k and node i.

x::Matrix{Int16}: solution x of the Lagrangian problem. x[k, a] gives the value of the solution x_a^k of the Lagrangian problem L(π) for demand k and arc a.

y::Vector{Int16}: solution y of the Lagrangian problem. y[a] gives the value of the solution y_a of the Lagrangian problem L(π) for arc a.

LRarcs::Vector{Float32}: Values of the Lagrangian problem for the arcs. LRarcs[a] gives the value of subproblem L_a associated with arc a.

objLR::Float32: Value of the Lagrangian dual problem

LearningPi.labelsCWL — Type

Label structure for the Bin Packing Problem. #Fields: -π: optimal lagrangian multipliers vector. -xLR: primal solution of the Knapsack Lagrangian Relaxation associated to the variables that associate one items to a pack (using the optimal Lagrangian multipliers). -yLR: primal solution of the Knapsack Lagrangian Relaxation associated to the variables say if we use or not a pack (using the optimal Lagrangian multipliers). -objLR: optimal value of the Lagrangian Dual

LearningPi.labelsGA — Type

Label structure for the Generalized Assignment Problem. #Fields: -π: optimal lagrangian multipliers vector. -xLR: primal solution of the Lagrangian Subproblem with optimal Lagrangian multipliers. -objLR: optimal value of the Lagrangian Dual.

LearningPi.learningType — Type

abstract learning type to type the functions that should work with all the type of models and features encoding

LearningPi.loss_hinge — Type

Structure of parameters for loss obtained as the inverse of the sub-problem obj value.

LearningPi.loss_hinge — Method

lossHinge(π; example, _::HingeLoss)

Loss function obtained taking the inverse of the sub-problem value for the predicted lagrangians.

Arguments:

π: lagrangian multipliers vector candidate.

example: dataset sample object.

-_: loss parameters, it should be a structure of type HingeLoss.

LearningPi.loss_mse — Method

loss_mse(π; example)

Arguments:

π: lagrangian multipliers vector candidate.

example: dataset sample object.

-_: loss parameters, it should be a structure of type MSELoss.

returns the loss function value obtained taking the MSE beteern the predicted Lagrangian multipliers π and the optimal ones in example.

LearningPi.lr_features_matrix — Type

struct lr_features_matrix

with this choice the features matrix considers the informations related to the continuous relaxation of the instance.

LearningPi.without_cr_features_matrix — Type

struct withou_cr_features_matrix

with this choice the features matrix does not considers the informations related to the continuous relaxation of the instance.

LearningPi.zero_deviation — Type

struct zero_deviation end

Type to use as deviation vector (i.e. the starting point from which our model produce an additive activation) the all zeros vector.

ChainRulesCore.rrule — Method

Compute the value of the Learning by Experience loss (usining the inverse of value of the sub-problem) and its pullback function.

ChainRulesCore.rrule — Method

Compute the value of the Learning by Experience loss (usining the inverse of value of the sub-problem) and its pullback function.

ChainRulesCore.rrule — Method

Compute the value of the Learning by Experience loss (usining the inverse of value of the sub-problem) and its pullback function.

LearningPi.compareWithBests — Method

function compareWithBests(currentMetrics::Dict,bestMetrics::Dict,nn,endString::String)

Arguments:

currentMetrics : a dictionary of Float

bestMetrics : a dictionary of Float

nn : a neural network

endString : a string used to memorize the best models

This function compare all the values in bestMetrics with the ones in currentMetrics (that corresponds to the same key). If some value in currentMetrics is better, then we update the correspondent value in bestMetrics and we save the model in a bson file.

LearningPi.createEmptyDataset — Method

createEmptyDataset(lt::learningGNN)

Arguments:

lt: learning Multi Layer Perceptron type.

Create an empty dataset for the Graph Neural network learning type.

LearningPi.createEmptyDataset — Method

createEmptyDataset(lt::learningMLP)

Arguments:

lt: learning Multi Layer Perceptron type.

Create an empty dataset for the Multi Layer Perceptron learning type.

LearningPi.create_features — Method

function create_features(ins::instanceGA)

#Arguments:
	- `ins`: instance object, it should be of type instanceGA 

read the features and returns a features structure.

LearningPi.create_features — Method

function create_features(ins::cpuInstanceMCND)

#Arguments:

ins: instance structure, should be of type cpuInstanceMCND

LearningPi.create_features — Method

function create_features(ins::instanceCWL)

#Arguments:
	- `ins`: instance object, it should be of type instanceCWL 

solve the Continuous Relaxation and the Lagrangian Sub-Problem considering as Lagrangian Multipliers
the dual variables associated to the relaxed constraints and then returns a features structure.

LearningPi.create_rms — Method

function create_rms(d)

#Arguments:

d: size of input in the input layers

return a RMSNorm function.

LearningPi.deviationFrom — Method

function deviationFrom(x,::crdeviation)

#Arguments:

x: the bipartite-graph representation of the instance

For the cr_deviation it returns the dual variables associated to the dualized constraints in the optimal solution of the continuous relaxation, taking the good components from the nodes features matrix in the bipartite-graph representation.

LearningPi.deviationFrom — Method

function deviationFrom(x,::zerodeviation)

#Arguments:

x: the bipartite-graph representation of the instance

For the zero_deviation it returns an all-zeros vector with the correct size. The size will be the same as the dual variables associated to the dualized constraints in the optimal solution of the continuous relaxation, taking the good components from the nodes features matrix in the bipartite-graph representation.

LearningPi.features_matrix — Method

function featuresmatrix(ins::instanceGA,featObj, fmt::abstractfeatures_matrix)

#Arguments:

ins: instance structure, it should be a sub-type of instanceGA

featObj: features object containing all the characteristics

fmt: features matrix type.

Construct the matrix of the features for a bipartite-graph representation of the instance.

LearningPi.features_matrix — Method

function featuresmatrix(ins::instanceMCND, featObj::featuresMCND, fmt::abstractfeatures_matrix)

#Arguments:

ins: instance structure, it should be a sub-type of abstractInstanceMCND

featObj: features object containing all the characteristics

fmt: features matrix type.

Construct the matrix of the features for a bipartite-graph representation of the instance.

LearningPi.features_matrix — Method

function featuresmatrix(ins::instanceCWL,featObj, fmt::abstractfeatures_matrix)

#Arguments:

ins: instance structure, it should be a sub-type of instanceCWL

featObj: features object containing all the characteristics

fmt: features matrix type.

Construct the matrix of the features for a bipartite-graph representation of the instance.

LearningPi.gap — Method

function gap(example::abstract_example, objPred::Real, objGold::Real, nInst::Real)

#Arguments:

example: the current example (dataset point).

objPred: the current obective for the example.

objGold: the optimal value of the Lagrangian Dual.

nInst: the number of the instances in the set.

compute the GAP of the instance in the example using the predicted objective objPred.

LearningPi.gap_closure — Method

function gap(example::abstract_example, objPred::Real, objGold::Real, nInst::Real)

#Arguments:

example: the current example (dataset point).

objPred: the current obective for the example.

objGold: the optimal value of the Lagrangian Dual.

nInst: the number of the instances in the set.

compute the closure GAP of the instance in the example using the predicted objective objPred. The closure is w.r.t. the value of the Lagrangian Sub-Problem, solved with the dual variables of the continuous relaxation.

LearningPi.get_cr_features — Method

function crfeatures(fmt::crfeatures_matrix, x, y)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

LearningPi.get_cr_features — Method

function crfeatures(fmt::lrfeatures_matrix, x, y)

#Arguments: -fmt: feature matrix type (it shoul be lrfeaturesmatrix)

LearningPi.get_cr_features — Method

function crfeatures(fmt::withoutcrfeaturesmatrix, _, _)

#Arguments: -fmt: feature matrix type (it shoul be withoutcrfeatures_matrix)

LearningPi.get_device — Method

function getdevice(los::abstractloss)

#Arguments:
- `_`: the loss parameters

returns the device (cpu/gpu) used to compute the loss.
For a general loss will be CPU.

LearningPi.get_model — Method

get_model(nn::GNNChain)

#Arguments: -nn: neural network model

returns a cpu version of the model that can be saved using a bson file.

LearningPi.get_parameters — Function

function get_parameters(nn, lt::learningType, f = 0)

nn : neural network

lt : learning type

f : ???

Returns the model parameters of nn in the case in which nn belsong to lt learning type.

LearningPi.get_λ — Method

function get_λ(x)

#Arguments:

x: the bipartite-graph representation of the instance

Returns the dual variables associated to the dualized constraints in the optimal solution of the continuous relaxation, taking the good components from the nodes features matrix in the bipartite-graph representation.

LearningPi.gradient_lsp — Method

function gradient_lsp( x, ins::cpuInstanceGA)

#Arguments:
- `x`: the solution of the Lagrangian Sub-problem
- `ins`: a cpuInstanceCWL structure

This function compute and returns the gradient of the sub-problem objective function w.r.t. the Lagrangian Multipliers.

LearningPi.gradient_lsp — Method

function gradient_lsp( x, ins::cpuInstanceCWL)

#Arguments:
- `x`: the solution of the Lagrangian Sub-problem
- `ins`: a cpuInstanceCWL structure

This function compute and returns the gradient of the sub-problem objective function w.r.t. the Lagrangian Multipliers.

LearningPi.gradient_lsp — Method

function gradient_lsp( x, ins::cpuInstanceMCND)

#Arguments:
- `x`: the solution of the Lagrangian Sub-problem
- `ins`: a cpuInstanceMCND structure

This function compute and returns the gradient of the sub-problem objective function w.r.t. the Lagrangian Multipliers.

LearningPi.printBests — Method

function printBests(bestMetrics::Dict, path::String)

Arguments:

bestMetrics: a dictionary of float that contains the best values

   find in the training for altypel the considered metrics.

path: location where print the results in a file.

Takes as input the dictionary of the best metrics and print the values in standard output and in a file defined by the path.

LearningPi.printMetrics — Method

function printMetrics(currentMetrics::Dict)

Arguments:

currentMetrics: a dictionary of Float

This function takes as input the dictionary of the metrics and print the values associated to training and validation sets.

LearningPi.print_best_models — Method

function printbestmodels(endString::String, bestModels::Dict)

#Arguments:

endString: Path where save the models

bestModels: Dictionary of the best models (w.r.t different metrics) found so far

Print in a file BSON, located in the folder endString the best model found so far

LearningPi.print_json — Method

function print_json(ins::cpuInstanceMCND, lab, feat, fileName,factory::cpuMCNDinstanceFactory)

#Arguments:

ins: instance structure, it should be of type CWLinstance

lab: labels structure, it should be of type labelsMCND

feat: features structure, it should be of type featuresMCND

fileName: the path to the file json where print the data

factory: instance factory should be of type cpuMCNDinstanceFactory

LearningPi.print_json — Method

function print_json(ins::instanceGA, lab::labelsGA, feat::featuresGA, fileName::String)

#Arguments: -ins: instance structure, it should be of type <: instanceGA

lab: labels structure, it should be of type labelsGA

feat: features structure, it should be of type featuresGA

fileName: the path to the file json where print the data

LearningPi.print_json — Method

function print_json(ins::instanceCWL, lab::labelsCWL, feat::featuresCWL, fileName::String)

#Arguments: -ins: instance structure, it should be of type <: instanceCWL

lab: labels structure, it should be of type labelsCWL

feat: features structure, it should be of type featuresCWL

fileName: the path to the file json where print the data

Print the information provided in the instance ins, the labels lab and the features feat in a JSON file located in the path fileName.

LearningPi.read_labels — Method

function read_labels(fileLabel::String, ins::instanceGA)

#Arguments:
	- `fileLabel`: the path to the file where to find labels informations
	- `ins`: instance object, it should be sub-type of instanceGA 

read the labels and returns a labels structure.

LearningPi.read_labels — Method

function read_labels(fileLabel::String, ins::abstractInstanceMCND)

#Arguments:
	- `fileLabel`: the path to the file where to find labels informations
	- `ins`: instance object, it should be of type abstractInstanceMCND 

read the labels and returns a labels structure.

LearningPi.read_labels — Method

function read_labels(fileLabel::String, ins::instanceCWL)

#Arguments:
	- `fileLabel`: the path to the file where to find labels informations
	- `ins`: instance object, it should be of type sub-type of instanceCWL 

read the labels and returns a labels structure.

LearningPi.saveHP — Method

function saveHP(endString::String,lr::Float32,decay::Float32,h::Vector{Int64},opt,lt::learningType,loss,seedNN::Int64,seedDS::Int64,stepSize::Int64)

Arguments:

endString: a string used as name for the output file.

lr: learning rate of the algorithm.

decay: decay for the learning rate.

h: a list of #(hidden layers), each component of h contains the number of nodes in

 the associated hidden layer.

opt: optimizer

lt: learning type object.

fmt:

dt:

loss: loss function.

seedDS: random seed for the dataset generation.

seedNN: random seed for the neural network parameters.

stepSize: the step size for the decay scheduler of the optimizer

nodes_number: size (number of nodes) in the hidden reprensentation between each layer

block_number: number of blocks in the model

hI: sizes of Dense layers in the first part, where the nodes features are sent in the hidden space

hF: sizes of Dense layers in the final

dataPath: path to the instances used in the dataset

factory: instance factory type

This function memorize all this hyper parameters in a JSON file.

LearningPi.saveHP — Method

function saveHP(endString::String,lr::Float32,decay::Float32,h::Vector{Int64}, opt, lt::learningType, loss,seedNN::Int64, seedDS::Int64, stepSize::Int64)

Arguments:

endString: a string used as name for the output file.

lr: learning rate of the algorithm.

decay: decay for the learning rate.

h: a list of #(hidden layers), each component of h contains the number of nodes in

 the associated hidden layer.

opt: optimizer

lt: learning type object.

loss: loss function.

seedDS: random seed for the dataset generation.

seedNN: random seed for the neural network parameters.

stepSize: the step size for the decay scheduler of the optimizer

This function memorize all this hyper parameters in a JSON file.

LearningPi.sizeFeatures — Method

sizeFeatures(lt::learningType,dS)

Arguments:

lt : learning type (general).

dS : dataset (corpus structure)

LearningPi.size_features_constraint — Method

function sizefeaturesconstraint(fmt::crfeaturesmatrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the constraints. In this case 6.

LearningPi.size_features_constraint — Method

function sizefeaturesconstraint(fmt::lrfeaturesmatrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the constraints. In this case 6.

LearningPi.size_features_constraint — Method

function sizefeaturesconstraint(fmt::withoutcrfeatures_matrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the constraints. In this case 4.

LearningPi.size_features_variable — Method

function sizefeaturesvariable(fmt::crfeaturesmatrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the variables. In this case 4.

LearningPi.size_features_variable — Method

function sizefeaturesvariable(fmt::lrfeaturesmatrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the variables. In this case 4.

LearningPi.size_features_variable — Method

function sizefeaturesvariable(fmt::withoutcrfeatures_matrix)

#Arguments: -fmt: feature matrix type (it shoul be crfeaturesmatrix)

returns the size of the features associated to the variables. In this case 2.

LearningPi.sub_problem_value — Method

function subproblemvalue(, v, example, _::abstractloss)

#Arguments:

_: lagrangian multipliers vector candidate,

v: the value of the loss function,

example: dataset sample object,

_: loss parameters.

Compute the value of the sub-problem for the loss for which it cannot be obtained in a smarter way.

LearningPi.sub_problem_value — Method

function subproblemvalue(_, v, example, _::HingeLoss)

#Arguments:

_: lagrangian multipliers vector candidate,

v: the value of the loss function,

example: dataset sample object,

_: loss parameters, it should be a structure of type HingeLoss.

Compute the value of the sub-problem without recomputing it, but using the value of the loss function (for the HingeLoss) and other informations contained in the sample

LearningPi.testAndPrint — Method

function testAndPrint(currentMetrics::Dict,testSet,nn,loss,loss,lt::learningType)

Arguments:

currentMetrics: a dictionary of Float

testSet : a vector of hingeExamples_instance that correspond to the test set.

nn : a neural network model.

loss: a structure with the parameters of the loss.

   For other details of the parameters of a certain loss see the definition of the particular structure of the loss.

lt: learning type object.

This function compute different metrics over the validation set. The values are memorized in the dictionary and print them in the standard output.

LearningPi.validation — Method

function validation(currentMetrics::Dict, valSet, nn,loss::abstractLoss, lt::learningType,dt::abstract_deviation)

Arguments:

currentMetrics: a dictionary of Float

valSet : a vector of gnn_dataset that correspond to the validation set.

nn : a neural network model.

loss: a structure with the parameters of the loss.

   For other details of the parameters of a certain loss see the definition of the particular structure of the loss.

loss: loss function.

lt: learning type object.

dt: deviation type (0 or dual of the CR)

This function compute different metrics over the validation set. The values are memorized in the dictionary.